services/text-extraction/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.35.0
kafka-python==2.0.2
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
loguru==0.7.2
